feat: add denyRequest function for error responses
